api-requirements.txt
flask-shell-ipython
PyJWT
SQLAlchemy
